fixed a bunch of warnings

+++ vigir_footstep_planning_default_plugins/src/robot_model/reachability_polygon.cpp
@@ -42,7 +42,7 @@ void ReachabilityPolygon::loadParams(const ParameterSet& params)
   step_range.clear();
   step_range.reserve(step_range_x.size());
 
-  for (size_t i = 0; i < step_range_x.size(); i++)
+  for (int i = 0; i < step_range_x.size(); i++)
   {
     double x = static_cast<double>(step_range_x[i]);
     double y = static_cast<double>(step_range_y[i]);
